rename of cancellation pulse protocol and adding hamiltonian tomography with amplitude sweeper (still work in progress thou, there might be some problems with the simulation for high amplitudes)
