+++ src/qibocal/protocols/two_qubit_interaction/cross_resonance/hamiltonian_tomography/cancellation_amplitude.py
@@ -7,7 +7,7 @@ channel with frequency set to the frequency of the target drive channel.
 
 import datetime
 from dataclasses import dataclass, field
-from typing import Any, Literal, Union
+from typing import Literal, Union
 
 import numpy as np
 import numpy.typing as npt
@@ -29,31 +29,30 @@ from .....auto.operation import (
     Routine,
 )
 from .....calibration import CalibrationPlatform
-from .....result import probability
 from ..utils import Basis, SetControl, cr_sequence
 from .utils import (
     EPS,
     HamiltonianTerm,
+    amplitude_tomography_cr_fit,
     calibration_cr_plot,
-    phase_tomography_cr_fit,
 )
 
-HamiltonianTomographyCRPhaseType = np.dtype(
+HamiltonianTomographyCANCAmplType = np.dtype(
     [
         ("prob_target", np.float64),
         ("error_target", np.float64),
         ("prob_control", np.float64),
         ("error_control", np.float64),
-        ("phase", np.float64),
+        ("amp", np.float64),
         ("x", np.int64),
     ]
 )
-"""Custom dtype for CR amplitude."""
+"""Custom dtype for Cancellation amplitude."""
 
 
 @dataclass
-class HamiltonianTomographyCRPhaseParameters(Parameters):
-    """HamiltonianTomographyCRAmplitude runcard inputs."""
+class HamiltonianTomographyCANCAmplParameters(Parameters):
+    """HamiltonianTomographyCANCAmplitude runcard inputs."""
 
     pulse_duration_start: float
     """Initial duration of CR pulse [ns]."""
@@ -61,14 +60,14 @@ class HamiltonianTomographyCRPhaseParameters(Parameters):
     """Final duration of CR pulse [ns]."""
     pulse_duration_step: float
     """Step CR pulse duration [ns]."""
-    control_phase_step: float
-    """Step CR pulse phase."""
-    control_phase: float = 0
-    """Initial phase of CR pulse."""
-    control_phase_end: float = 2 * np.pi
-    """Final phase of CR pulse."""
-    control_amplitude: Union[dict[QubitPairId, float], float] = 0.1
+    control_amplitude: float
     """Initial amplitude of CR pulse."""
+    target_amplitude: float
+    """Amplitude of cancellation pulse."""
+    target_ampl_end: float
+    """Final amplitude of CR pulse."""
+    target_ampl_step: float
+    """Step CR pulse amplitude."""
     interpolated_sweeper: bool = False
     """Use real-time interpolation if supported by instruments."""
     echo: bool = False
@@ -78,12 +77,12 @@ class HamiltonianTomographyCRPhaseParameters(Parameters):
     """
 
     @property
-    def phase_range(self) -> np.ndarray:
+    def amplitude_range(self) -> np.ndarray:
         """Amplitude range for CR pulses."""
         return np.arange(
-            self.control_phase,
-            self.control_phase_end,
-            self.control_phase_step,
+            self.target_amplitude,
+            self.target_ampl_end,
+            self.target_ampl_step,
         )
 
     @property
@@ -95,21 +94,19 @@ class HamiltonianTomographyCRPhaseParameters(Parameters):
 
 
 @dataclass
-class HamiltonianTomographyCRPhaseResults(Results):
-    """HamiltonianTomographyCRAmplitude outputs."""
+class HamiltonianTomographyCANCAmplResults(Results):
+    """HamiltonianTomographyCANCAmpl outputs."""
 
     echo: bool
-
     control_amplitude: Union[dict[QubitPairId, float], float]
-
     hamiltonian_terms: dict[QubitPairId, dict] = field(default_factory=dict)
     """Terms in effective Hamiltonian."""
     fitted_parameters: dict[tuple[QubitId, QubitId], dict[HamiltonianTerm, list]] = (
         field(default_factory=dict)
     )
-    """Fitted parameters from Hamiltonian Terms values for different phases."""
+    """Fitted parameters for Hamiltonian Terms values for different amplitudes."""
 
-    cancellation_pulse_phases: dict[QubitPairId, dict[str, float]] = field(
+    cancellation_pulse_amplitudes: dict[QubitPairId, dict[str, float]] = field(
         default_factory=dict
     )
     """Fitted parameters for cancellation pulse phases."""
@@ -121,37 +118,39 @@ class HamiltonianTomographyCRPhaseResults(Results):
 
 
 @dataclass
-class HamiltonianTomographyCRPhaseData(Data):
+class HamiltonianTomographyCANCAmplData(Data):
     """Data structure for CR Amplitude."""
 
     echo: bool
-    control_amplitude: Any = (None,)
-    phases: list | None = None
+    control_amplitude: Union[dict[QubitPairId, float], float]
+    amplitudes: list | None = None
     data: dict[
         tuple[QubitId, QubitId, Basis, SetControl],
-        npt.NDArray[HamiltonianTomographyCRPhaseType],
+        npt.NDArray[HamiltonianTomographyCANCAmplType],
     ] = field(default_factory=dict)
+    """Raw data acquired."""
 
     @property
     def pairs(self):
         return {(i[0], i[1]) for i in self.data}
 
-    def select_phase(self, phase: float):
-        new_data = HamiltonianTomographyCRPhaseData(
+    def select_amplitude(self, amplitude: float):
+        new_data = HamiltonianTomographyCANCAmplData(
             echo=self.echo,
+            control_amplitude=self.control_amplitude,
         )
-        new_data.data = {k: d[d.phase == phase] for k, d in self.data.items()}
+        new_data.data = {k: d[d.amp == amplitude] for k, d in self.data.items()}
         return new_data
 
     def register_qubit(self, dtype, data_keys, data_dict):
         """Store output for single qubit."""
         duration_list = data_dict["x"]
-        phase_list = data_dict["phase"]
-        size = len(duration_list) * len(phase_list)
+        amp_list = data_dict["amp"]
+        size = len(duration_list) * len(amp_list)
         ar = np.empty(size, dtype=dtype)
-        phases, durations = np.meshgrid(phase_list, duration_list)
+        amplitudes, durations = np.meshgrid(amp_list, duration_list)
         ar["x"] = durations.ravel()
-        ar["phase"] = phases.ravel()
+        ar["amp"] = amplitudes.ravel()
         ar["prob_target"] = data_dict["prob_target"].ravel()
         ar["error_target"] = data_dict["error_target"].ravel()
         ar["prob_control"] = data_dict["prob_control"].ravel()
@@ -161,11 +160,11 @@ class HamiltonianTomographyCRPhaseData(Data):
 
 
 def _acquisition(
-    params: HamiltonianTomographyCRPhaseParameters,
+    params: HamiltonianTomographyCANCAmplParameters,
     platform: CalibrationPlatform,
     targets: list[QubitPairId],
-) -> HamiltonianTomographyCRPhaseData:
-    """Data acquisition for Hamiltonian tomography CR ctrl_phaseprotocol.
+) -> HamiltonianTomographyCANCAmplData:
+    """Data acquisition for Hamiltonian tomography CR protocol.
 
     We measure the expectation values X,Y and Z on the target qubit after
     applying the CR sequence specified by the input parameters. We repeat the
@@ -176,9 +175,9 @@ def _acquisition(
 
     """
 
-    data = HamiltonianTomographyCRPhaseData(
+    data = HamiltonianTomographyCANCAmplData(
         echo=params.echo,
-        phases=params.phase_range.astype(float).tolist(),
+        amplitudes=params.amplitude_range.astype(float).tolist(),
         control_amplitude=params.control_amplitude,
     )
 
@@ -187,24 +186,20 @@ def _acquisition(
 
         for basis in Basis:
             for setup in SetControl:
-                # if basis == Basis.Z and setup==SetControl.Id:
                 sequence, cr_pulses, cr_target_pulses, delays = cr_sequence(
                     platform=platform,
                     control=control,
                     target=target,
-                    amplitude=(
-                        params.control_amplitude[pair]
-                        if isinstance(params.control_amplitude, dict)
-                        else params.control_amplitude
-                    ),
-                    phase=params.control_phase_end,
-                    target_amplitude=0,
-                    target_phase=0,
+                    amplitude=params.control_amplitude,
+                    phase=None,
+                    target_amplitude=params.target_ampl_end,
+                    target_phase=None,
                     duration=params.pulse_duration_end,
                     echo=params.echo,
                     setup=setup,
                     basis=basis,
                 )
+
                 if params.interpolated_sweeper:
                     length_sweeper = Sweeper(
                         parameter=Parameter.duration_interpolated,
@@ -217,21 +212,12 @@ def _acquisition(
                         values=params.duration_range,
                         pulses=cr_pulses + cr_target_pulses + delays,
                     )
-                phase_sweepers = [
-                    Sweeper(
-                        parameter=Parameter.relative_phase,
-                        values=params.phase_range,
-                        pulses=[cr_pulses[0]],
-                    )
-                ]
-                if params.echo:
-                    phase_sweepers.append(
-                        Sweeper(
-                            parameter=Parameter.relative_phase,
-                            values=params.phase_range + np.pi,
-                            pulses=[cr_pulses[1]],
-                        )
-                    )
+
+                amp_sweeper = Sweeper(
+                    parameter=Parameter.amplitude,
+                    values=params.amplitude_range,
+                    pulses=cr_target_pulses,
+                )
 
                 updates = []
                 updates.append(
@@ -243,14 +229,13 @@ def _acquisition(
                         }
                     }
                 )
-                acquisition_type = AcquisitionType.INTEGRATION
-                # acquisition_type = AcquisitionType.DISCRIMINATION
+
                 results = platform.execute(
                     [sequence],
-                    [[length_sweeper], phase_sweepers],
+                    [[length_sweeper], [amp_sweeper]],
                     nshots=params.nshots,
                     relaxation_time=params.relaxation_time,
-                    acquisition_type=acquisition_type,
+                    acquisition_type=AcquisitionType.DISCRIMINATION,
                     averaging_mode=AveragingMode.CYCLIC,
                     updates=updates,
                 )
@@ -261,19 +246,15 @@ def _acquisition(
                     sequence.channel(platform.qubits[control].acquisition)
                 )[-1].id
 
-                if acquisition_type == AcquisitionType.INTEGRATION:
-                    prob_target = results[target_acq_handle]
-                    prob_control = results[control_acq_handle]
-                else:
-                    prob_target = probability(results[target_acq_handle], state=1)
-                    prob_control = probability(results[control_acq_handle], state=1)
+                prob_target = results[target_acq_handle].ravel()
+                prob_control = results[control_acq_handle].ravel()
 
                 data.register_qubit(
-                    HamiltonianTomographyCRPhaseType,
+                    HamiltonianTomographyCANCAmplType,
                     (control, target, basis, setup),
                     dict(
                         x=length_sweeper.values,
-                        phase=phase_sweepers[0].values,
+                        amp=amp_sweeper.values,
                         prob_target=1 - 2 * prob_target,
                         error_target=2
                         * np.sqrt(
@@ -287,52 +268,54 @@ def _acquisition(
                 )
 
     t = datetime.datetime.now().strftime("%d_%m_%Y_%H:%M:%S")
-    np.savez(f"./{t}_acquisition_data_phase_tomography", data)
+    np.savez(f"./{t}_acquisition_data_amplitude_tomography", data)
 
     return data
 
 
 def _fit(
-    data: HamiltonianTomographyCRPhaseData,
-) -> HamiltonianTomographyCRPhaseResults:
-    """Post-processing function for HamiltonianTomographyCRAmplitude.
+    data: HamiltonianTomographyCANCAmplData,
+) -> HamiltonianTomographyCANCAmplResults:
+    """Post-processing function for HamiltonianTomographyCANCAmpl.
 
     We fit the expectation values using the Eq. S10 from the paper https://arxiv.org/pdf/2303.01427.
     Afterwards, we extract the Hamiltonian terms from the fitted parameters.
 
     """
-    hamiltonian_terms, fitted_parameters, pulses_phases = phase_tomography_cr_fit(
+    hamiltonian_terms, fitted_parameters, canc_amplitudes = amplitude_tomography_cr_fit(
         data=data,
     )
 
-    return HamiltonianTomographyCRPhaseResults(
+    return HamiltonianTomographyCANCAmplResults(
         echo=data.echo,
         control_amplitude=data.control_amplitude,
         hamiltonian_terms=hamiltonian_terms,
         fitted_parameters=fitted_parameters,
-        cancellation_pulse_phases=pulses_phases,
+        cancellation_pulse_amplitudes=canc_amplitudes,
     )
 
 
 def _plot(
-    data: HamiltonianTomographyCRPhaseData,
+    data: HamiltonianTomographyCANCAmplData,
     target: QubitPairId,
-    fit: HamiltonianTomographyCRPhaseResults,
+    fit: HamiltonianTomographyCANCAmplResults,
 ):
-    """Plotting function for HamiltonianTomographyCRAmplitude."""
+    """Plotting function for HamiltonianTomographyCANCAmpl."""
     figs, fitting_report = calibration_cr_plot(data, target, fit)
     return figs, fitting_report
 
 
 def _update(
-    results: HamiltonianTomographyCRPhaseResults,
+    results: HamiltonianTomographyCANCAmplResults,
     platform: CalibrationPlatform,
     target: QubitPairId,
 ):
     # here is updated the full CNOT Pulse Sequence, which is composed by a CR sequence followe by a X_pi/2 and Z_(-pi/2) rotations on
     # target and control qubit respectively
 
-    target = target[::-1] if target not in results.cancellation_pulse_phases else target
+    target = (
+        target[::-1] if target not in results.cancellation_pulse_amplitudes else target
+    )
 
     new_cr_seq, _, _, _ = cr_sequence(
         platform=platform,
@@ -342,9 +325,9 @@ def _update(
         if isinstance(results.control_amplitude, dict)
         else results.control_amplitude,
         duration=None,
-        phase=results.cancellation_pulse_phases[target]["control"],
-        target_amplitude=0,
-        target_phase=results.cancellation_pulse_phases[target]["target"],
+        phase=None,
+        target_amplitude=results.cancellation_pulse_amplitudes[target]["ampl_iy"],
+        target_phase=None,
         echo=results.echo,
         setup=SetControl.Id,
         basis=Basis.Y,
@@ -357,5 +340,5 @@ def _update(
     getattr(update, f"{results.native}_sequence")(new_cr_seq, platform, target)
 
 
-hamiltonian_tomography_cr_phase = Routine(_acquisition, _fit, _plot, _update)
-"""HamiltonianTomography Routine object."""
+hamiltonian_tomography_canc_amplitude = Routine(_acquisition, _fit, _plot, _update)
+"""HamiltonianTomographyCANCAmpl Routine object."""
